backend/requirements.txt
# Flask Core
Flask==3.1.2
Werkzeug==3.1.3

# WebSocket Support
Flask-SocketIO==5.5.1
eventlet==0.39.0

# CORS
Flask-CORS==6.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
alembic==1.14.0

# Environment Variables
python-dotenv==1.0.1

# Type Checking
mypy==1.14.1

# Code Quality
pylint==3.3.6

# LLM Integration
ollama==0.6.0

# Symbolic Math
sympy==1.14.0

# Production Server
gunicorn==23.0.0
